Refactor Ui.java to interfaces.ui pckg
