+++ src/main/java/uxie/Uxie.java
@@ -6,7 +6,7 @@ import uxie.exceptions.UxieIOException;
 import uxie.interfaces.CommandParse;
 import uxie.interfaces.Storage;
 import uxie.interfaces.TaskList;
-import uxie.interfaces.Ui;
+import uxie.interfaces.ui.Ui;
 
 /**
  * Main class of Chatbot Uxie.
