+++ src/main/java/uxie/commands/DeleteCommand.java
@@ -4,7 +4,7 @@ import uxie.exceptions.UxieIOException;
 import uxie.exceptions.UxieIllegalOpException;
 import uxie.interfaces.Storage;
 import uxie.interfaces.TaskList;
-import uxie.interfaces.Ui;
+import uxie.interfaces.ui.Ui;
 
 /**
  * Command for deleting a task.
