+++ src/main/java/uxie/commands/ExitCommand.java
@@ -2,7 +2,7 @@ package uxie.commands;
 
 import uxie.interfaces.Storage;
 import uxie.interfaces.TaskList;
-import uxie.interfaces.Ui;
+import uxie.interfaces.ui.Ui;
 
 /**
  * Command that exits Uxie program.
