+++ src/main/java/uxie/commands/FindCommand.java
@@ -5,7 +5,7 @@ import java.util.List;
 import uxie.exceptions.UxieIllegalOpException;
 import uxie.interfaces.Storage;
 import uxie.interfaces.TaskList;
-import uxie.interfaces.Ui;
+import uxie.interfaces.ui.Ui;
 
 /**
  * Command that finds Tasks matching String.
