+++ src/main/java/uxie/commands/TodoCommand.java
@@ -3,7 +3,7 @@ package uxie.commands;
 import uxie.exceptions.UxieIOException;
 import uxie.interfaces.Storage;
 import uxie.interfaces.TaskList;
-import uxie.interfaces.Ui;
+import uxie.interfaces.ui.Ui;
 import uxie.tasks.ToDo;
 
 /**
