+++ src/main/java/uxie/commands/UnmarkCommand.java
@@ -4,7 +4,7 @@ import uxie.exceptions.UxieIOException;
 import uxie.exceptions.UxieIllegalOpException;
 import uxie.interfaces.Storage;
 import uxie.interfaces.TaskList;
-import uxie.interfaces.Ui;
+import uxie.interfaces.ui.Ui;
 
 /**
  * Command that marks a task as incomplete.
