+++ src/main/java/uxie/interfaces/ui/Ui.java
@@ -1,4 +1,4 @@
-package uxie.interfaces;
+package uxie.interfaces.ui;
 
 import java.util.Scanner;
